+++ logic/src/main/java/ua/com/tracksee/logic/facade/OrderFacade.java
@@ -12,6 +12,7 @@ import ua.com.tracksee.logic.driver.DriverOrderBean;
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.HashMap;
 import java.util.List;
@@ -213,15 +214,15 @@ public class OrderFacade {
         return driverOrderBean.getHistoryOfOrders(id, pageNumber);
     }
 
-    public int getOrdersPagesCountCompleted(int id){
+    public BigInteger getOrdersPagesCountCompleted(int id){
         return driverOrderBean.getOrdersPagesCountCompleted(id);
     }
 
-    public int getOrdersPagesCountQueued(UserEntity driver){
+    public BigInteger getOrdersPagesCountQueued(UserEntity driver){
         return driverOrderBean.getOrdersPagesCountQueued(driver);
     }
 
-    public int getOrdersPagesCountAssigned(int id){
+    public BigInteger getOrdersPagesCountAssigned(int id){
         return driverOrderBean.getOrdersPagesCountAssigned(id);
     }
